fix all 4.03 warnings in the codebase
- warning 45 was raised in canopy_store.ml as a local open introduces
  Ok/Error, which now shadow the standard `result` type
  constructors. The `let open .. in` is turned into a `let module
  C = .. in` to work around this. See also
  [MPR#7386](http://caml.inria.fr/mantis/view.php?id=7386).

- an unused identifier warning was fixed

- `Omd.of_string content |> Omd.to_html` warns as `to_html` has
  a ton of implicit parameters that are "implicitly omitted" in
  this higher-order style. I'm not very happy about it, but the
  dumb rewrite into `Omd.to_html (Omd.of_string content)` avoids
  the warning.

This commit is in fact incomplete as there are instances of warning 42
(which warns on *any* use of type-based disambiguation) left in the
codebase. Canopy itself cannot do anything about these instances of
disambiguation, they come from the fact that Syndic, in version 1.5.1,
has two distinct Text constructors in Syndic.Atom -- I guess this was
not the case before.

I fixed it locally by editing the Makefile to add -42 to warn(..), but
I don't know how to change it in config.ml. Any help?