fix created/update dates from branches
In f4b6454476b2d193f31bcc0b29ba15e2a882f68b we introduced support for
fetching the blog content from a branch, but the code to compute
creation and update dates for individual post is still assuming that
the 'master' branch should be used by default; as a result, blog posts
would appear in the Atom feed in the wrong order. This seems to be due
to a discrepancy in the code style that does not occur anywhere else
in the codebase. The discrepancy is now fixed and the Atom feed is as
expected.

Note that the Atom feed is still ordered by creation date instead of
last-update date, which may or may not correspond to authors'
expectation. Either are fine for my current use-case.